Added new helper folder && added auto-include for those
Added new git_helper and sh_helper
