+++ lib/fastlane.rb
@@ -21,6 +21,7 @@ module Fastlane
   Helper = FastlaneCore::Helper # you gotta love Ruby: Helper.* should use the Helper class contained in FastlaneCore
 
   Fastlane::Actions.load_default_actions
+  Fastlane::Actions.load_helpers
 
   if Fastlane::FastlaneFolder.path
     actions_path = File.join(Fastlane::FastlaneFolder.path, 'actions')
