+++ lib/fastlane/actions/actions_helper.rb
@@ -8,30 +8,6 @@ module Fastlane
       ENVIRONMENT = :ENVIRONMENT
     end
 
-    # Helper Methods
-    def self.git_author
-      s = `git log --name-status HEAD^..HEAD`
-      s = s.match(/Author:.*<(.*)>/)[1]
-      return s if s.to_s.length > 0
-      return nil
-    rescue
-      return nil
-    end
-
-    def self.last_git_commit
-      s = `git log -1 --pretty=%B`.strip
-      return s if s.to_s.length > 0
-      nil
-    end
-
-    # Returns the current git branch - can be replaced using the environment variable `GIT_BRANCH`
-    def self.git_branch
-      return ENV['GIT_BRANCH'] if ENV['GIT_BRANCH'].to_s.length > 0 # set by Jenkins
-      s = `git rev-parse --abbrev-ref HEAD`
-      return s.to_s.strip if s.to_s.length > 0
-      nil
-    end
-
     def self.executed_actions
       @executed_actions ||= []
     end
@@ -78,62 +54,24 @@ module Fastlane
       raise exc if exc
     end
 
-    # Execute a shell command
-    # This method will output the string and execute it
-    # Just an alias for sh_no_action
-    # @param log [boolean] should fastlane print out the executed command
-    def self.sh(command, log: true)
-      sh_no_action(command, log: log)
-    end
-
-    def self.sh_no_action(command, log: true)
-      # Set the encoding first, the user might have set it wrong
-      previous_encoding = [Encoding.default_external, Encoding.default_internal]
-      Encoding.default_external = Encoding::UTF_8
-      Encoding.default_internal = Encoding::UTF_8
-
-      command = command.join(' ') if command.kind_of?(Array) # since it's an array of one element when running from the Fastfile
-      Helper.log.info ['[SHELL COMMAND]', command.yellow].join(': ') if log
-
-      result = ''
-      if Helper.test?
-        result << command # only for the tests
-      else
-        exit_status = nil
-        IO.popen(command, err: [:child, :out]) do |io|
-          io.each do |line|
-            Helper.log.info ['[SHELL]', line.strip].join(': ')
-            result << line
-          end
-          io.close
-          exit_status = $?.exitstatus
-        end
-
-        if exit_status != 0
-          # this will also append the output to the exception
-          raise "Exit status of command '#{command}' was #{exit_status} instead of 0. \n#{result}"
-        end
-      end
-
-      result
-    rescue => ex
-      raise ex
-    ensure
-      Encoding.default_external = previous_encoding.first
-      Encoding.default_internal = previous_encoding.last
-    end
-
     # returns a list of official integrations
     # rubocop:disable Style/AccessorMethodName
     def self.get_all_official_actions
-      Dir[File.expand_path '*.rb', File.dirname(__FILE__)].collect do |file|
+      Dir[File.expand_path('*.rb', File.dirname(__FILE__))].collect do |file|
         File.basename(file).gsub('.rb', '').to_sym
       end
     end
     # rubocop:enable Style/AccessorMethodName
 
     def self.load_default_actions
-      Dir[File.expand_path '*.rb', File.dirname(__FILE__)].each do |file|
+      Dir[File.expand_path('*.rb', File.dirname(__FILE__))].each do |file|
+        require file
+      end
+    end
+
+    # Import all the helpers
+    def self.load_helpers
+      Dir[File.expand_path('../helper/*.rb', File.dirname(__FILE__))].each do |file|
         require file
       end
     end
